+++ packages/clerk-js/src/core/resources/Base.ts
@@ -127,10 +127,18 @@ export abstract class BaseResource {
 
       assertProductionKeysOnDev(status, errors);
 
-      throw new ClerkAPIResponseError(message || statusText, {
-        data: errors,
-        status: status,
-      });
+      const apiResponseOptions: ConstructorParameters<typeof ClerkAPIResponseError>[1] = { data: errors, status };
+      if (status === 429 && headers) {
+        const retryAfter = headers.get('retry-after');
+        if (retryAfter) {
+          const value = parseInt(retryAfter, 10);
+          if (!isNaN(value)) {
+            apiResponseOptions.retryAfter = value;
+          }
+        }
+      }
+
+      throw new ClerkAPIResponseError(message || statusText, apiResponseOptions);
     }
 
     return null;