+++ packages/clerk-js/src/ui/polishedAppearance.ts
@@ -232,12 +232,14 @@ export const polishedAppearance: Appearance = {
       card: {
         ...cardContentStyles(theme),
       },
-      planCardDefault: {
-        borderWidth: 0,
-        boxShadow: `${theme.shadows.$cardBoxShadow}, ${BORDER_SHADOW_LENGTH} ${theme.colors.$neutralAlpha100}`,
-      },
-      planCardCompact: {
-        boxShadow: 'none',
+      pricingTableCard: {
+        '&[data-variant="default"]': {
+          borderWidth: 0,
+          boxShadow: `${theme.shadows.$cardBoxShadow}, ${BORDER_SHADOW_LENGTH} ${theme.colors.$neutralAlpha100}`,
+        },
+        '&[data-variant="compact"]': {
+          boxShadow: 'none',
+        },
       },
       scrollBox: {
         ...cardContentStyles(theme),